requirements.txt
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Django==5.1.3
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
tzdata==2024.2
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
